move dataclass
