+++ antarest/study/business/study_interface.py
@@ -18,9 +18,10 @@ from typing_extensions import override
 
 from antarest.core.exceptions import CommandApplicationError
 from antarest.matrixstore.service import ISimpleMatrixService
-from antarest.study.dao.api.study_dao import ReadOnlyStudyDao, StudyMetadata
+from antarest.study.dao.api.study_dao import ReadOnlyStudyDao
 from antarest.study.dao.file.file_study_dao import FileStudyTreeDao
 from antarest.study.dao.memory.in_memory_study_dao import InMemoryStudyDao
+from antarest.study.model import StudyMetadata
 from antarest.study.storage.rawstudy.model.filesystem.factory import FileStudy
 from antarest.study.storage.variantstudy.model.command.common import CommandOutput
 from antarest.study.storage.variantstudy.model.command.icommand import ICommand
